Add files via upload
This MATLAB code simulates a random conformation of a chromosome arm
in a yeast cell's nucleus. This script is intended for any researcher who 
wishes to learn how to model a chromosome in a simple manner. It is also
useful for teaching college (or advanced high school) students how to
create a Monte Carlo simulation.