Add files via upload
This MATLAB code simulates a random conformation of the arm of a chromosome 
in a yeast cell's nucleus. This script is intended for any researcher who 
wishes to model a chromosome in a simple manner and wants to reduce the
computation time compared to doing a regular random walk chromosome simulation.
This code uses the FAST ALGORITHM: This algorithm speeds up 
the simulation of a freely-jointed chain chromosome that is confined to a 
certain region and/or attached to certain locations. This algorithm was 
developed by Bronk for the paper "Chromosome-refolding model of mating-type
switching in yeast" published in PNAS in 2016 by Avaroglu, Bronk, et al.
The algorithm is described in the paper's Supplementary Information under
the heading "Polymer Model Simulation Details."